teams_job/requirements.txt
google-cloud-bigquery==3.14.0
requests==2.32.3